requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==41.0.1
cssbeautifier==1.14.8
defusedxml==0.7.1
Django==3.2.19
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-multiupload==0.6.1
django-widget-tweaks==1.4.12
djlint==1.30.2
EditorConfig==0.12.3
frozenlist==1.3.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
install==1.3.5
jsbeautifier==1.14.8
json5==0.9.14
magic-filter==1.0.9
multidict==6.0.4
oauthlib==3.2.2
pathspec==0.11.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
